Fix for selection of default Azure account
`az account show` and `armClient.GetDefaultSubscription` don't agree on what default azure subscription is. As a result, the tool could show one subscription in the title but the resources belonging to another subscription.
With the fix, the resources belonging to the subscription listed by `az account show` will be listed.

I also sorted the resources alphabetically, like in Azure portal.